Move Take to its own file
